refactor RouteDetailVC/ add routeDetailView
