+++ RideTogether/Controller/RouteList/RouteDetailViewController.swift
@@ -13,8 +13,6 @@ class RouteDetailViewController: UIViewController {
     
     var routeData: RouteData!
     
-    @IBOutlet weak var routeImageView: UIImageView!
-    
     @IBOutlet weak var routeMapView: MKMapView! {
         
         didSet {
@@ -23,23 +21,15 @@ class RouteDetailViewController: UIViewController {
         }
     }
     
-    @IBOutlet weak var mapBackView: UIView! {
+    @IBOutlet weak var routeDetailView: RouteDetailView! {
         
         didSet {
             
-            mapBackView.addRoundOnTop()
+            routeDetailView.addRoundOnTop()
             
-            mapBackView.addShadow(offset: CGSize(width: 3, height: -2), opacity: 0.4)
+            routeDetailView.addShadow(offset: CGSize(width: 3, height: -2), opacity: 0.4)
         }
     }
-    
-    @IBOutlet weak var distanceLabel: UILabel!
-    
-    @IBOutlet weak var elevationLabel: UILabel!
-    
-    @IBOutlet weak var locationLabel: UILabel!
-    
-    @IBOutlet weak var aveAlopeLabel: UILabel!
     
     @IBAction func backLastPage() {
         
@@ -58,29 +48,14 @@ class RouteDetailViewController: UIViewController {
     
     func setupInform() {
         
+        routeDetailView.setView(routeData: routeData)
+        
         let polylineCode = routeData.map.polyline
         
         PolylineManager.shared.mapView = self.routeMapView
         
         PolylineManager.shared.showPolyline(polylineCode: polylineCode)
-        
-        let urlString = "https://bike100.tw/wp-content/uploads/\(routeData.routeID).jpg"
-        
-        let aveAlope: Double = {
-            
-            return (routeData.distance / 1000) / (routeData.elevationGain / 100)
-        }()
-        
-        locationLabel.text = routeData.name
-        
-        routeImageView.setImage(urlString: urlString)
-        
-        distanceLabel.text = String(format: "%.2f", (routeData.distance) * 2 / 1000) + " km"
-        
-        elevationLabel.text = String(format: "%.2f", routeData.elevationGain) + " m"
-        
-        aveAlopeLabel.text = String(format: "%.1f", aveAlope) + " %"
-    }
+     }
 }
 
 extension RouteDetailViewController: MKMapViewDelegate {
